Time: 30 ms (11.88%), Space: 5.9 MB (92.38%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int mySqrt(int x) 
     {
-        double comp{1};
+        long long comp{1};
         
         while(comp * comp <= x)
         {
